# Conflicts:
#   Controllers/PatientsController.cs

+++ Controllers/ProceduresController.cs
@@ -23,17 +23,25 @@ namespace apka2.Controllers
         // GET: Procedures
         public async Task<IActionResult> Index()
         {
-            if (getSessionUserId() == 0)
+            var doctorId = getSessionUserId();
+            if (doctorId == 0)
             {
                 return RedirectToAction("AccessDenied", "Doctors");
             }
-            return View(await _context.Procedure.ToListAsync());
+            if (getIsAdmin() == 1)
+            {
+                return View(await _context.Procedure.ToListAsync());
+            }
+
+            var surveysIds = getSurveysIds(doctorId);
+            return View(_context.Procedure.Where(p => surveysIds.Contains(p.SurveyId)));
         }
 
         // GET: Procedures/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (getSessionUserId() == 0)
+            var doctorId = getSessionUserId();
+            if (doctorId == 0)
             {
                 return RedirectToAction("AccessDenied", "Doctors");
             }
@@ -42,14 +50,17 @@ namespace apka2.Controllers
                 return NotFound();
             }
 
-            var procedure = await _context.Procedure
-                .FirstOrDefaultAsync(m => m.Id == id);
+            var surveysIds = getSurveysIds(doctorId);
+            var procedure = getIsAdmin() == 1 ? await _context.Procedure
+                .FirstOrDefaultAsync(p => p.Id == id)
+                : await _context.Procedure
+                .FirstOrDefaultAsync(p => p.Id == id && surveysIds.Contains(p.SurveyId));
             if (procedure == null)
             {
                 return NotFound();
             }
             ViewData["sessions"] = _context.ProcedureSession.
-                Where(m => m.ProcedureId == id);
+                Where(ps => ps.ProcedureId == id);
             return View(procedure);
         }
 
@@ -167,7 +178,8 @@ namespace apka2.Controllers
         // GET: Procedures/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (getSessionUserId() == 0)
+            var doctorId = getSessionUserId();
+            if (doctorId == 0)
             {
                 return RedirectToAction("AccessDenied", "Doctors");
             }
@@ -181,6 +193,13 @@ namespace apka2.Controllers
             {
                 return NotFound();
             }
+
+            var surveysIds = getSurveysIds(doctorId);
+            if (getIsAdmin() == 0 && surveysIds.Contains(procedure.SurveyId))
+            {
+                return RedirectToAction("AccessDenied", "Doctors");
+            }
+
             return View(procedure);
         }
 
@@ -191,7 +210,8 @@ namespace apka2.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,WasEnded,Anticoagulation,SurveyId,ProcedureDate,ECMO,Filter,ProcedureTime,ExtracorporealClearingMethod,CitrateConcentrate,UnplanedTermination,TerminationReason,BloodReturn,PatientDeath,DeathDate,Remarks")] Procedure procedure)
         {
-            if (getSessionUserId() == 0)
+            var surveysIds = getSurveysIds(getSessionUserId());
+            if (getIsAdmin() == 0 && surveysIds.Contains(procedure.SurveyId))
             {
                 return RedirectToAction("AccessDenied", "Doctors");
             }
@@ -226,7 +246,8 @@ namespace apka2.Controllers
         // GET: Procedures/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (getSessionUserId() == 0)
+            var doctorId = getSessionUserId();
+            if (doctorId == 0)
             {
                 return RedirectToAction("AccessDenied", "Doctors");
             }
@@ -242,6 +263,12 @@ namespace apka2.Controllers
                 return NotFound();
             }
 
+            var surveysIds = getSurveysIds(doctorId);
+            if (getIsAdmin() == 0 && surveysIds.Contains(procedure.SurveyId))
+            {
+                return RedirectToAction("AccessDenied", "Doctors");
+            }
+
             ViewData["sessions"] = _context.ProcedureSession.
                 Where(m => m.ProcedureId == id);
 
@@ -253,7 +280,8 @@ namespace apka2.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (getSessionUserId() == 0)
+            var doctorId = getSessionUserId();
+            if (doctorId == 0)
             {
                 return RedirectToAction("AccessDenied", "Doctors");
             }
@@ -264,6 +292,11 @@ namespace apka2.Controllers
             var procedure = await _context.Procedure.FindAsync(id);
             if (procedure != null)
             {
+                var surveysIds = getSurveysIds(doctorId);
+                if (getIsAdmin() == 0 && surveysIds.Contains(procedure.SurveyId))
+                {
+                    return RedirectToAction("AccessDenied", "Doctors");
+                }
                 _context.Procedure.Remove(procedure);
 
                 var sessions = _context.ProcedureSession.
@@ -292,6 +325,28 @@ namespace apka2.Controllers
                 return 0;
             }
             return (int)sessionId;
+        }
+
+        private int getIsAdmin()
+        {
+            var isAdmin = HttpContext.Session.GetInt32(SessionData.SessionKeyIsAdmin);
+            if (isAdmin == null)
+            {
+                return 0;
+            }
+            return (int)isAdmin;
+        }
+
+        private List<int> getSurveysIds(int doctorId)
+        {
+            var surveys = _context.Survey.
+                Where(s => s.Id == doctorId).ToList();
+            var surveysIds = new List<int>();
+            foreach (var survey in surveys)
+            {
+                surveysIds.Add(survey.Id);
+            }
+            return surveysIds;
         }
     }
 }
